fix: tighten pagination coercion to integer-only with .int()
Appends `.int()` to the three `z.coerce.number()` calls in
`PaginationSchema` (`page`, `pageSize`, `totalCount`), making the
integer intent explicit and rejecting non-integer floats such as
`"1.5"` that would otherwise coerce silently to a float.  Adds a
regression test asserting that a `"1.5"` page value is rejected.

The string-to-integer coercion path for well-formed pagination values
(e.g. `"1"`, `"20"`) is unchanged and all prior tests continue to pass.

+++ SparkyFitnessServer/schemas/foodSchemas.ts
@@ -58,10 +58,11 @@ export const PaginationSchema = z.object({
   // these pagination values as strings, and the same field can switch between a
   // string and a number across requests. Coerce them so any provider that sends
   // string-typed numeric pagination is normalized rather than failing response
-  // validation. `hasMore` stays a strict boolean.
-  page: z.coerce.number(),
-  pageSize: z.coerce.number(),
-  totalCount: z.coerce.number(),
+  // validation. `.int()` makes the integer intent explicit and rejects
+  // non-integer floats (e.g. "1.5"). `hasMore` stays a strict boolean.
+  page: z.coerce.number().int(),
+  pageSize: z.coerce.number().int(),
+  totalCount: z.coerce.number().int(),
   hasMore: z.boolean(),
 });
